upgraded to latest ormlite

+++ yafsrc/ServiceStack.OrmLite/src/ServiceStack.Interfaces/Redis/Generic/IRedisSortedSet.Generic.cs
@@ -1,38 +1,38 @@
 //
-// https://github.com/mythz/ServiceStack.Redis
+// https://github.com/ServiceStack/ServiceStack.Redis
 // ServiceStack.Redis: ECMA CLI Binding to the Redis key-value storage system
 //
 // Authors:
 //   Demis Bellot (demis.bellot@gmail.com)
 //
-// Copyright 2010 Liquidbit Ltd.
+// Copyright 2014 Service Stack LLC. All Rights Reserved.
 //
-// Licensed under the same terms of Redis and ServiceStack: new BSD license.
+// Licensed under the same terms of ServiceStack.
 //
 
 using System.Collections.Generic;
-using ServiceStack.DesignPatterns.Model;
+using ServiceStack.Model;
 
 namespace ServiceStack.Redis.Generic
 {
-	public interface IRedisSortedSet<T> : ICollection<T>, IHasStringId
-	{
-		T PopItemWithHighestScore();
-		T PopItemWithLowestScore();
-		double IncrementItem(T item, double incrementBy);
-		int IndexOf(T item);
-		int IndexOfDescending(T item);
-		List<T> GetAll();
-		List<T> GetAllDescending();
-		List<T> GetRange(int fromRank, int toRank);
-		List<T> GetRangeByLowestScore(double fromScore, double toScore);
-		List<T> GetRangeByLowestScore(double fromScore, double toScore, int? skip, int? take);
-		List<T> GetRangeByHighestScore(double fromScore, double toScore);
-		List<T> GetRangeByHighestScore(double fromScore, double toScore, int? skip, int? take);
-		int RemoveRange(int minRank, int maxRank);
-		int RemoveRangeByScore(double fromScore, double toScore);
-		double GetItemScore(T item);
-		int PopulateWithIntersectOf(params IRedisSortedSet<T>[] setIds);
-		int PopulateWithUnionOf(params IRedisSortedSet<T>[] setIds);
-	}
+    public interface IRedisSortedSet<T> : ICollection<T>, IHasStringId
+    {
+        T PopItemWithHighestScore();
+        T PopItemWithLowestScore();
+        double IncrementItem(T item, double incrementBy);
+        int IndexOf(T item);
+        long IndexOfDescending(T item);
+        List<T> GetAll();
+        List<T> GetAllDescending();
+        List<T> GetRange(int fromRank, int toRank);
+        List<T> GetRangeByLowestScore(double fromScore, double toScore);
+        List<T> GetRangeByLowestScore(double fromScore, double toScore, int? skip, int? take);
+        List<T> GetRangeByHighestScore(double fromScore, double toScore);
+        List<T> GetRangeByHighestScore(double fromScore, double toScore, int? skip, int? take);
+        long RemoveRange(int minRank, int maxRank);
+        long RemoveRangeByScore(double fromScore, double toScore);
+        double GetItemScore(T item);
+        long PopulateWithIntersectOf(params IRedisSortedSet<T>[] setIds);
+        long PopulateWithUnionOf(params IRedisSortedSet<T>[] setIds);
+    }
 }
