+++ yafsrc/ServiceStack.OrmLite/src/ServiceStack.Interfaces/Redis/IRedisHash.cs
@@ -1,13 +1,13 @@
 using System.Collections.Generic;
-using ServiceStack.DesignPatterns.Model;
+using ServiceStack.Model;
 
 namespace ServiceStack.Redis
 {
-	public interface IRedisHash
-		: IDictionary<string, string>, IHasStringId
-	{
-		bool AddIfNotExists(KeyValuePair<string, string> item);
-		void AddRange(IEnumerable<KeyValuePair<string, string>> items);
-		int IncrementValue(string key, int incrementBy);
-	}
+    public interface IRedisHash
+        : IDictionary<string, string>, IHasStringId
+    {
+        bool AddIfNotExists(KeyValuePair<string, string> item);
+        void AddRange(IEnumerable<KeyValuePair<string, string>> items);
+        long IncrementValue(string key, int incrementBy);
+    }
 }
